Lodash, QR code
