+++ components/common/ScanQRModal/ScanQRButton.tsx
@@ -1,9 +1,8 @@
-import React, { useState, Suspense, type ReactElement } from 'react'
-import dynamic from 'next/dynamic'
+import React, { lazy, useState, Suspense, type ReactElement } from 'react'
 import QrCodeIcon from '@mui/icons-material/QrCode'
 import { IconButton } from '@mui/material'
 
-const ScanQRModal = dynamic(() => import('.'))
+const ScanQRModal = lazy(() => import('.'))
 
 type Props = {
   onScan: (value: string) => void
